near: getFinalizedBlockNumber works
