+++ watcher/package.json
@@ -25,6 +25,7 @@
     "dotenv": "^16.0.3",
     "firebase-admin": "^11.4.0",
     "js-sha512": "^0.8.0",
+    "near-api-js": "^1.1.0",
     "winston": "^3.8.2"
   },
   "devDependencies": {
